# Backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CancelWorkOrderRequest:
    properties:
      reason:
        example: Customer cancelled the order
        type: string
    required:
    - reason
    type: object
...
  handlers.CompleteWorkOrderRequest:
    properties:
      actual_cost:
        example: 950
        type: number
      completion_notes:
        example: Work completed successfully
        type: string
    required:
    - actual_cost
    - completion_notes
    type: object
...
  handlers.ErrorResponse:
    description: Standard error response model
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.SuccessResponse:
    description: Standard success response model
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.WorkOrderItemInfo:
    properties:
      description:
        example: Fabric material
        type: string
      id:
        example: 1
        type: integer
      item_id:
        example: 1
        type: integer
      quantity:
        example: 100
        type: integer
      total_price:
        example: 1000
        type: number
      unit_price:
        example: 10
        type: number
    type: object
  handlers.WorkOrderItemInput:
    properties:
      description:
        example: Fabric material
        type: string
      item_id:
        example: 1
        type: integer
      quantity:
        example: 100
        type: integer
      unit_price:
        example: 10
        type: number
    required:
    - item_id
    - quantity
    - unit_price
    type: object
  handlers.WorkOrderResponse:
    properties:
      actual_cost:
        example: 950
        type: number
      assigned_to:
        example: 1
        type: integer
      completion_notes:
        example: Completed on time
        type: string
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      customer_name:
        example: John Doe
        type: string
      description:
        example: Production of 100 jerseys
        type: string
      end_date:
        example: "2024-03-26T21:41:49Z"
        type: string
      estimated_cost:
        example: 1000
        type: number
      id:
        example: 1
        type: integer
      items:
        items:
          $ref: '#/definitions/handlers.WorkOrderItemInfo'
        type: array
      order_id:
        example: 1
        type: integer
      spk_number:
        example: SPK-2024-001
        type: string
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: draft
        type: string
      tasks:
        items:
          $ref: '#/definitions/handlers.WorkOrderTaskItem'
        type: array
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
      work_type:
        example: production
        type: string
    type: object
...
  handlers.WorkOrderTaskItem:
    properties:
      assigned_to:
        example: 1
        type: integer
      description:
        example: Cut the fabric according to patterns
        type: string
      end_date:
        example: "2024-03-24T23:41:49Z"
        type: string
      id:
        example: 1
        type: integer
      notes:
        example: Ready for cutting
        type: string
      start_date:
        example: "2024-03-24T21:41:49Z"
        type: string
      status:
        example: pending
        type: string
      task_name:
        example: Cutting
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Vomo backend API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vomo API
  version: "1.0"
paths:
...
  /work-orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Mark a work order as cancelled with a reason
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cancellation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CancelWorkOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkOrderResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a work order
      tags:
      - WorkOrder
  /work-orders/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a work order as completed with completion notes and actual
        cost
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Completion Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CompleteWorkOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkOrderResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete a work order
      tags:
      - WorkOrder
  /work-orders/{id}/items:
    post:
      consumes:
      - application/json
      description: Create a new item for a work order
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item Data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkOrderItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.WorkOrderItemInfo'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new work order item
      tags:
      - WorkOrder
  /work-orders/{id}/items/{itemId}:
    delete:
      description: Delete an item from a work order
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order or item not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a work order item
      tags:
      - WorkOrder
    put:
      consumes:
      - application/json
      description: Update an existing item in a work order
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Work Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Item Data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkOrderItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkOrderItemInfo'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Work order or item not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a work order item
      tags:
      - WorkOrder
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"